Input string should be shorted in lexicographical order
The returned string should have 26 characters and the characters in the input should be present in it. If the there is no character that is present in the string then replace with '.' in the resultant string.

Consider only the alphabetic characters. Ignore the special characters and if repeated character occurs ignore it too.